corecfg: also "mask" services when disabling them
This is a port of https://github.com/snapcore/core/pull/61:

In addition to disabling services when
snap set service.{ssh,syslog}.disable=true is called we need to
mask them. The reason is that the /etc/systemd/system directory
is marked "synced" and the core snap has a "syslog.service" symlink.
So on each boot if there is no "syslog.service" (or ssh) link there
writable-path will happily copy it. This is not what we want. By
using "mask" in addition to disable systemd will create a symlink
with the name of the service to /dev/null. This prevents
writable-path from interfering.