snap-confine: use #include in snap-confine.apparmor.in
This unbreaks the python apparmor tools that do not support the
"include" keyword. Tools like aa-enforce currently fail on stable
because of our use of include. While this is arguable an apparmor
python tools bug we still need to help fix the situation by
changing our code to use "#include" which is supported.