hookstate: add compat "configure-snapd" task.
We created a `configure-snapd` task for some versions of snapd
between 2.29 and 2.30. We do not use this task anymore but systems
that generated it will not refresh properly because the task is
unknown in the latest snapd. Add a compatibility wrapper for this
task. We can remove that at some point in the future.