tests: make less calls to the package manager
instead of, e.g.,

    for p in "${pkgs[@]}"; do apt install "$p"; done

this does

    apt install "${pkgs[@]}"

(but also for dnf, etc). This should be quicker and slightly less
susceptible to network issues in the metadata querying phase of dnf.
Allegedly.

Also (mostly because this is a resurfacing of an old commit), use gzip
instead of xz for mksquashfs in tests a little more (should be
quicker).