tests/main/postrm-purge: stop snapd before purge and reinstall the package in restore
The 'purge' functionality is to be run when removing the package from the
system. Make sure that snapd services are stopped, like they would be when the
step was done through the package manager. Cleanup the remaining package files
before the test finishes.

Reinstall the package in restore to provide a consistent 'environment' for
subsequent tests.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>