tests/main/interfaces-fuse_support: refactor the test
Refactor the test to start the fuse-consumer in foreground and make it
background job, thus avoiding the need for listing all processes and
cherry-picking the pids.

Make sure we ask fuse-consume to do a clean exit and verify its exit status and
then verify that mount point was removed from the namespace.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>