snap: improve error for snaps not available in the given context
When we have a snap in the store that is not available in the
channel that the user requested we currently show a "not-found"
error. But it is better to display more meaningful information.

This is a RFC PR that adds this but has the cost of one extra
round-trip to the store.

Fixes https://bugs.launchpad.net/snapstore/+bug/1665787