cmd/libsnap-confine-private: print failed mount/umount command when debugging is enabled
When running with debugging enabled, the mount/umount command will not be
printed to stderr unless the build was done with -DSNAP_CONFINE_DEBUG_BUILD.
However, we'd still want to see the command if the operation fails. The patch
fixes the behavior, and does not overwrite the [u]mount_cmd helper variable.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>