cmd: fix snap-confine-debug, SNAP_CONFINE_DEBUG_BUILD does not propagate to libsnap-confine-private
snap-confine-debug is a 'debug' build of snap-confine, thus the
SNAP_CONFINE_DEBUG_BUILD is defined in CFLAGS. However, some bits that are
enabled by SNAP_CONFINE_DEBUG_BUILD are in the source code of
libsnap-confine-private.a and setting a CFLAGS on snap-confine-debug binary does
not propagate to the lib. Work around this by pulling in the sources of
libsnap-confine-private.a into snap-confine-debug build.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>