cmd: fix snap-confine-debug, SNAP_CONFINE_DEBUG_BUILD does not propagate to libsnap-confine-private
snap-confine-debug is a 'debug' build of snap-confine, hence
SNAP_CONFINE_DEBUG_BUILD flag is set in CFLAGS. However, some bits that are
enabled by SNAP_CONFINE_DEBUG_BUILD are in the source code of
libsnap-confine-private.a and setting a CFLAGS on snap-confine-debug binary does
not propagate to the lib. Work around this by pulling in the debug build of
libsnap-confine-private.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>