cmd/snap-update-ns: refactor Change.Perform
This patch is my attempt at refactoring the code to include support for
bind mounting directories as well as bind mounting files, and creating
symlinks.

I tried to reuse the heavily tested SecureMk{dir,file}All functions but
chose to remove the ensureMountPoint helper as it was insufficient. For
extra clarity in the critical part I chose to roll all of the code into
Change.Perform, with appropriate amount of explanation as to what is
going on.

The existing tests are working fine. I will follow up with extra tests
for the new code paths that touch files and symlink specific logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>