cmd/snap-update-ns: adjust inaccurate comment
Bind mounts to files are perfectly usable as long as apps just open
the file in question (and not attempt to move it or rename it) but as much
is expected of directory bind mounts so this has the same set of
limitations. As such, they are supported (they were by the code already
but the comment was stale).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>