cmd/snap-update-ns: refactor and improve Change.Perform to handle EROFS
This patch makes the Change.Perform function correctly handle a number
of elements, collectively being able to pierce writable holes in read
only filesystems. This builds upon the prior work on planning and
construction of a writable mimic that uses tmpfs and a bind mount/symlink
farm to reconstruct the content of the original directory.

The details are interesting: existing directories and files are
bind mounted onto empty directories / files specifically created for
this purpose. For simplicity files are annotated with "x-snapd.kind=file"
mount option. Directories remain the default for backwards compatibility
and thus don't use a new mount option. Symlinks are different as they
need to be directly re-created as actual symlinks (symlinks do not
participate mounts directly).

As a part of the refactor the ensureMountPoint function was discarded
and inlined into changePerform. This was necessary as there are slightly
different responsibilities necessary for each of the cases that are
also compounded by the need to retry after constructing writable space.

This patch is augmented with a separate large refactor of unit tests,
for readability and clarity but to ensure that this is not breaking
anything I will propose it separately, showing that existing unit tests
don't break.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>