snap: add `snap run --strace` to be able to strace snap apps
This is a first (naive) implementation to support
`snap run --strace`. It is based on the work by Jamie outlined
in https://forum.snapcraft.io/t/1433

It will use the strace from outside to snap.

This has the ugly side-effect that the strace output will include
all the things that snap-confine and snap-run are doing. It also
means that the command is run via sudo (but strace will drop to
the right user). It will also try to use the "strace-static" snap
if there is no system strace.

No support for strace options yet, but easy to add if the approach
is considered worth persuing.

An alternative approach is to run strace much later - from
`snap-run`. However this means we are inside the confinement
already so we need to make sure that all the syscalls that
strace needs are allowed by the confinement (and also that
apparmor works). We could bind-mount the strace-static command
but we still need a way to modify the confinement which always
has the risk that the traced app suddently behaves differently
than without strace. Yet another idea would be to stop the
snap command in `snap run` by some magic and then use the
`strace -p` feature to attach to the pid of the child of
`snap run` and then resume the operation of the child of
snap-run. Ideas welcome!