tests: update "searching" test to match store changes
The current searching test is broken because there is no
databases section anymore. The test is now updated and
uses "featured" as the section which is hopefully stable.

It also switches to search for the "vlc" snap in the video
section which hopefully stays stable for some time (it is
a very popular snap so hopefully it won't get pulled/renamed
any time soon).