many: fix removal of snaps inside LXD
On LXD the root filesystem is not mounted with sharing of mount events
so we don't get propagation of mount and unmount events, as snaps get
added and removed, into persistent mount namespaces.

To work around this ship a snap.mount systemd unit that creates a
bind-mount at /snap and makes it shared. The unit is only active inside
LXD which is a high-priority target.

We may want to expand on this to other containers supported in the same
way but this will be in another patch. One of the existing tests that
checked for the sharing is now removed as it artificially broke what we
were doing, instead a new test, explicitly using LXD is added, that
doesn't do any magic manipulation and should be representative of
real-world use.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>