cmd/snap-confine: warn about missing sharing opts
When /snap (or equivalent) is not sharing mount events then we will run
in an environment where the main mount namespace will correctly function
but per-snap mount namespaces and future per-user mount namespaces will
not see any changes made outside (snaps being mounted and unmounted) and
will eventually cease to function correctly as the initial revision goes
away.

Ideally we'd be loud and fail early but since we didn't before (we just
tried to fix it, incorrectly) let's start by emitting a gentle warning.
Over time that warning should be transformed into a snapd system
inspection where snapd can actively detect that the environment is not
supported (or that there's a bug).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>