snap: improve `snap run` robustness
Before an empty cmdRun{} would run strace by default because
cmdRun.Strace is "" and that was assumed to be meant that we
got `snap run --strace` (without further args). However that
is confusing in e.g. main.go where we use cmdRun{} to run a
snap command. By adding a new cmdRun.ParserRan bool and checking
for that in cmdRun.useStrace we can avoid this confusing behavior.