overlord/configstate/config: make SetSnapConfig delete on empty
With the previous SetSnapConfig it wasn't possible to pass in an empty
(nil or []byte{}) raw config; the serialisation would fail. It also
wasn't possible to unset a config; at best you could set it to `{}`.

This addresses both issues: a nil or empty config can now be used to
clear the config.

This makes using it very natural, as when the config isn't set,
GetSnapConfig will return a nil (raw) config as well. The caller can
treat the returned raw config as an opaque blob, as intended.