cmd/snap-confine: allow constructing layouts (phase 1)
This patch implements the first phase of layouts agreed with jdstrand.

The requirements for this phase are:
 * source of mount --bind must be in on of $SNAP, $SNAP_DATA or $SNAP_COMMON
 * target of symlink must in in one of $SNAP, $SNAP_DATA, or $SNAP_COMMON
 * may not mount on top of an existing layout mountpoint

The first and rules are implemented with an apparmor profile. All three
rules are implemented with additional snapd-side validation. As an extra
sanity check mounts over /media are denied as that directory is setup
with bidirectional mount event propagation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>