advisor: ensure commands.db has mode 0644 and add test
When running bolt.Open() even in readonly mode the DB file will
be created if it is missing. Our read-only advisor.Open() uses
mode 0600 to open the Db in read-only mode. This means that if
a root user triggers command-not-found and the DB file is not
yet downloaded the DB is created with 0600. This happend during
manual testing. This PR switches permissions to 0644. It also
adds a test to ensure that the file has the correct permissions.