osutil: allow using many globs in EnsureDirState
This patch simply extends the directory synchronizer to allow more than
one glob to be used. This will come in handy when we want to provide
more than one kind of content for a given snap and then the
prefix/suffix (that is expressed as a glob) cannot capture that if only
one glob is available.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>