Update kill-timeout focused on make tests pass on boards
This is happening that the prepare phase takes about 2.5 to 3 minutes in
a dragonboard in case the sd card is not fast enough. This is making the
test suite fail during the prepare phase when we test the new core on
edge channel and beta channel.

The idea is to move from 3m to 5m the kill-timeout to avoid this kind of
issues.

Example: https://paste.ubuntu.com/26506907/