Change the root path from tmp to /home/test/tmp to avoid lack of space issue
This change is because in some boards like the dragonboard there is not
enough space in /tmp to unpack the snaps. Because of that, the
kernel.img is not in $IMAGE/boot/uboot/pi2-kernel_49.snap/ making the
test prepare-image-uboot fail.

It is still missing a research about why the command "snap prepare-
image" is not failing and making the test fail when the board goes out
of space.