daemon: remove redundant UserOK markings from api commands
We currently have a bunch of Commands in the API that are using
`UserOK:` in a redundant (and for my eyes confusing) way.

The default is `UserOK: false` so we don't need to state it. It
also only applies to `GET` so `UserOK: true` for something that
does not have a `GET` has no effect.

The PR also documents what the canAccess() policy.