snap: add abstraction layer for layout validation
Layout validation is somewhat tricky. To make it more scalable and
easier to understand layout elements are now validated against a list of
constraints that can be specialized to particular layout entries
(mounted file-systems, symlinks, etc.)

In addition some more tests show how this behaves in practice. Some of
the ideas are thanks to Chipaca, thank you man!

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>