debian, snap: only static link libseccomp in snap-seccomp on ubuntu
Rework how we do static linking of libseccomp for the snap-seccomp
helper. The most recent golang change for 1.9 added a whitelist
of allowed things for #cgo LDFLAGS (and #cgo PKG_CONFIG) which
broke out current approach. It seems the only way to get back the
static libseccomp linking is via the CGO_LDFLAGS_ALLOW environment.
So use that during the package build. For the other distros this
is less of an issue, we need the static libseccomp on Ubuntu as
snap-seccomp goes into the core snap and is called if snapd is
re-execing. This means that we cannot have a dynamic libseccomp
there or we might get into libseccomp library compatibility
issues between the host and the core.