cmd/snap-update-ns: don't ignore layout construction errors
When a layout cannot be constructed then don't just log the error and
carry on. Fail loudly and early on. This is done so that applications
can rely on layout application to work and won't have to consider any
partial states.

In addition this will fix any security issues involved with additional
read and write permissions resulting from layouts. For example if
/etc/shadow is supposed to be $SNAP/etc/shadow but the layout
construction fails then the snap would still have write access to
/etc/shadow. This patch fixes it by refusing to run the snap.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>