timeutil: fix scheduling on nth weekday of the month
The patch fixes scheduling on nth weekdays of the month. In particular checking
if the date being tested really is the nth occurrence of given weekday weekday
and whether if properly falls into range between nth weekdays.

Consider Feb 2017:

       February 2017
   Su Mo Tu We Th Fr Sa
             1  2  3  4
    5  6  7  8  9 10 11
   12 13 14 15 16 17 18
   19 20 21 22 23 24 25
   26 27 28

Schedule 'mon1-tue2' would incorrectly include dates 01-05.02, because they are
in the 1st week of the month, but not do not meet the schedule as they are not
'after' the first Monday of the month. Also dates 07-14.02 would be incorrectly
excluded as they are 'after' Tuesday in the 2nd week of the month, while in fact
the dates fall within the range of the schedule as they precede the 2nd Tuesday
of the month.

Address this by changing how the week-span checks are performed. In detail,
replace a simple week-count check with actually finding the nth occurrence of
given weekday in a month,

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>