interfaces/network: allow AF_CONN for userspace SCTP
Userspace SCTP (usrsctp) defines the AF_CONN domain as used with socket(). As
a result, the protocol is not known to the kernel and is not mediated by
AppArmor (usrsctp is encapsulated in protocols that are mediated). As such, we
define AF_CONN in snap-seccomp and add it to the network plug. Add a note to
the default AppArmor template for why we don't list it there.

References:
https://github.com/sctplab/usrsctp/blob/master/usrsctplib/usrsctp.h