update snap-confine and default profile for / on overlayfs
Currently (as of 4.13) when using overlayfs with AppArmor, we need directory
rules for the overlayfs upperdir. We add /upper/ rules to snap-confine and
the default template when '/' is overlayfs (ie, '/upper/...'). This is
compatible with so called 'livecd' environments which are setup like:

  /cow on / type overlay (...,lowerdir=...,upperdir=/cow/upper,workdir=...)

We may choose to add other paths or detection of upperdir at a later date.

References:
https://launchpad.net/bugs/1703674