tests/lib/prepare-restore: sync journal before rotating and vacuuming
Attempt to sync the journal before rotating. This should cause all in-flight
logs to be be written to the backing store. Thus when we rotate and vacuum, the
removed journals will have all the logs from previous test runs and the 'old'
logs should not be present during current test run. This is important if the
test is using the same snap with services as the previous one.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>