many: drop snaps' InstallDate, introduce Updated
Before this change, snaps would have an `InstallDate` attribute
in the client API (`"install-date"` in the JSON). The field would
not be present for 'remote' snaps (so, not in `/v2/find`
results), only in 'local' (`/v2/snaps`) snaps, and it was wrong:
the value returned was the timestmap of the toplevel directory of
the snap, which is not the install date of the snap.

The *actual* install date is also probably confusing to users, as
the date it was installed could be interpreted as the date the
user did `snap install`, or as of was when _this revision_ was
installed (FWIW it was meant to be the latter). In `snap info`,
for example, we call this time `refreshed` time (oblivious to the
fact that it was wrong much of the time).

As far as I know (I checked in 16.04's GNOME's software; maybe
@ikeydoherty and @robert-ancell can confirm) GUI software stores
aren't using the timestamp; snapd-glib will return `NULL` for
snapd_snap_get_install_date, and because in searches it won't be
there clients will have had to handle that `NULL`.

Between it not always being there, it being wrong when it is, and
it being called an ambiguous thing even if it weren't, I think
the best thing we can do is just drop it.

This change drops `InstallDate`, and adds `Updated` in its place.

* for 'remote' snaps, `Updated` is populated from the store's
  `LastUpdated` field.

* for 'local' snaps, `Updated` is populated from the timestamp of
  the thing we put in `/var/lib/snapd/snaps` (that'd be the
  `.snap` for regular snaps, or a symlink to the thing for `try`
  snaps -- so we use `lstat`, fwiw). It being called something
  different from `Refreshed` avoids it causing confusion when
  it's found to be different from timestamps in `snap changes`.

This will still leave a small element of possible confusion, in
that `Updated` will not update on revert, but hopefully this is
minor (and if it isn't, needs addressing separately; it isn't a
problem introduced by this change).

There is also another timestamp that would be quite useful to
report, for local squashfs snaps, and that's the one of the
squash's creation. This might follow in a separate PR.

> #### open question:
> In both cases, `snap info` will list this as `refreshed`, which
> is a little weird for remote snaps (and leaves open the
> possibility of confusion with what `snap changes` reports). I
> could change it to say `updated`, but I don't know how well
> that'd fly.