many: simplify mocking of home-on-NFS
The function that determines if /home is on an NFS filesystem is
provided by the osutil module and is tested there. We don't need to mock
the data it uses to compute the result but rather the result itself.
This simplifies mocking, removes unneeded, exported mock functions,
removes redundant tests and shortens the code overall.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>