daemon, snap: fix InstallDate, make a method of *snap.Info
(client.Snap).InstallDate was being set from the timestmap of the
toplevel directory of a snap, instead of from the timestamp of the
current symlink.

This is probably a direct consequence of it being tucked away in a
helper function in daemon/snap.go. I've made it more visible by making
it a method on snap.Info.