cmd/snap-update-ns: use syscall.Symlink instead of os.Symlink
We are handling and expecting raw syscall errors, not  the wrapped
variety. This fixes the in-the-wild ability to handle read-only medium
with symlinks. The existing unit tests mocked the return value of
syscall.Symlink, so unit tests did not catch this.

With this patch integration tests now almost passes (it is stopped by
another bug that is fixed separately).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>