cmd/snap-update-ns: switch to recursive directory bind mounts
This patch fixes a bug where poking writable holes in places with
existing bind mounts would "undo" those bind mounts because only the
first layer would be "copied" by a bind mount. Recursive bind mount
will transfer all of the directoriy and file bind mounts over.

With this change we must also switch to lazy unmount (MNT_DETACH) on
the undo operation. This is required because otherwise we would have
to iterate the mount table and figure out which mount entries to
remove. Using MNT_DETACH simply makes the kernel do that for us.

To convey the new unmount flag we now have "x-snapd.detach" mount option
on some generated unmount changes.

The apparmor profile for snap-update-ns has been updated to include
the new rbind permissions (both bind and rbind are now required).
In addition, testing this has uncovered that we didn't have enough
permissions to undo the writable mimic plan in case of failures
(we lacked appropriate umount permission). This is now fixed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>