cmd/snap: simple autostart for `snap userd`
Add autostart mode to `snap userd` helper, that can be invoked by running `snap
userd --autostart`. In this mode `userd` will attempt to autostart user
applications, by listing the desktop files in ~/snap/*/current/.config/autostart
across all snaps.

We use a simple heuristic, the autostart application desktop file must be named
the same as the snap app declared in snap.yaml. The application is started using
the command from a desktop file that was autogenerated when the snap was
installed.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>