configstate: when disable "ssh" we must disable the "sshd" service
When `snap set core service.ssh.disable=true` is run we must disable
the sshd.service instead of the ssh.service. The ssh.service is
an alias for sshd, but if ssh.service is disabled we create a
"mask" file on /etc/systemd/ssh.service. But we need a mask file
for "sshd.service" because the "sshd.service" file is also in the
core snap and the "writable-path" magic will bring it back after
a reboot.

LP: 1752031