snap-seccomp: conditionally use ActLog for devmode
Allow some distros to make use of the ActLog action for devmode, despite
the upstream libseccomp-golang project not yet cutting a release, by
guessing at the const value. The guess is validated by looking at the
string representation of the action to make sure that it is the expected
string for ActLog.

This is possible because the mvo5 github fork of libseccomp-golang
contains the minimal changes necessary to support ActLog. The system
must also have a new enough libseccomp and kernel to support the log
action.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>