wrappers: refactor desktop file sanitizer to support autostart files
Refactor the code for sanitizing and rewriting desktop files to accomodate
autostart files. These files are dropped by various applications in
XDG_CONFIG_HOME/autostart for the purpose of enabling their startup when the
desktop session is brought up.

Autostart desktop file is matched with a given snap application if that
application's command is present in the Exec=.. line and the application was
declared to use a desktop file with that name in snap.yaml.

The rules for cleaning up *.desktop files shipped in meta/gui/ remain the same.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>