progress: tweak ansimeter cvvis use to no longer confuse minicom
minicom's vt220 implementation was improperly displaying half of
cvvis. It seems it can't handle CPI codes separated by ;, and fails to
ignore the unknown escape.

This simplifies it, but it still does what we want (the cvvis we were
using, taken from terminfo for xterm, was additionally doing an 'echo
on', which we don't need).

Additionally I received feedback about the spinner being nastier than
the old slash-based one, so I moved to that.

... some day I might make that customisable :-)