osutil: DropPrivs morphs an *exec.Cmd to drop privileges
This is not our old friend “drop privileges in Go using Setuid”; this is its
humbler sibling, “drop privileges using sudo when running a command”.

Just in case that's too cryptic, this PR also uses `osutil.DropPrivs` to drop
privileges in `snap/squashfs`, so you can see it in action.

The immediate driver for this is that dropping to an unprivileged user
before running unsquashfs to e.g. read the snap.yaml will let us
postpone hashing the file without worrying about that attack vector.

It resurrects the user "snappypkg" for this. It's not a very nice name, but
has the advantage of already existing in core.

This'll need tweaks to all the packaging flavours. Patches welcome...