cmd/snap-apparmor-parser: add a prototype apparmor parser
This patch adds a small wrapper around the real apparmor parser that
handles caching in a sane way. The input is a source apparmor text. The
text is pre-processed to resolve all include statements. The
pre-processed text is cached with SHA1. The resulting hash is used as a
file name in /var/cache/snapd/apparmor. The compilation is only done if
the cache item is missing. This handles changing of the input text and
any included (#include) files. At no time does the cache use timestamps
for validity.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>