errtracker: respect the /etc/whoopsie configuration
We must support the settings of the user about sending problem
reports. This is implemented on Ubuntu via the /etc/whoopsie
configuration and that config is controllable via the
gnome-control-center.

We should probably also have our own internal configuration
(snapd.problem_reports={true,false}?). But that will be a
followup PR.