run-checks: handle 1.10's new! improved! inanity.
In go 1.10 when there are no tests to run a cover profile is still
generated, containing only the `mode:` line. This makes our
`append_coverage` function fail, which fails the whole suite.

I fixed `append_coverage` to not fail.

While I was there and thinking of 1.10, I added code so that, if
running with go 1.10, and given that 1.10 can do coverage of multiple
packages simultaneously, it just does the whole suite at once. This
doesn't use `append_coverage` at all.

Belts and suspenders all the way down.