cmd/snap-update-ns: fix creation of layout symlinks
This patch addresses two issues with layout symlinks:

 - symlinks were not created in a secure way, because symlinkat(2) and
   readlinkat(2) were not available
 - attempting to create a symlink where one would be present and had
   the right target would fail

This addresses the first half of the layout feature that sometimes
breaks in spread tests because existing symbolic links are not
inspected.

As a small tweak the secureMklinkAll function was renamed to
secureMksymlinkAll. In case we ever want to create hard links the name
will be less confusing.

The new function is covered with some unit tests and some non-mocked
real tests, ensuring that it is fully tested without extensive (long)
set of mocked unit tests.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>