strutil, cmd/snap: drop strutil.WordWrap, first pass at replacement
strutil.WordWrap was rather buggy, so it's getting replaced.

This is the first pass at that, which is already good enough that I
think this should land, but it does have more work to be done. Gustavo
would like to find a heuristic for line folding; I'd like to make it
more aware of unicode widths.

In any case once we're happy with it we should probably move the
reusable parts to strutil.