cmd/snap-update-ns: don't fail on existing symlinks
There's one more path where we check if a symlink is in place and this
path was not updated to not fail on existing but correct symlinks.

While I'm working on a simplification that drops all of this needless
code I thought about making a quick patch that can land in the next
release safely.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>