tests: make autopkgtest tests more targeted
The autopkgtest tests run in a different environment/lifecycle than
most of the other spread backends. An autopkgtest may run weeks
after the actual release as part of a reverse dependency test for
a different package that snapd uses. For this reason we exclude
some tests have external inputs (like the store search test).

The autopkgtest environment is also slower and more restricted than
our regular spread machines. This means that some expensive tests
(like lxd or core transition) should not be run there as this may
lead to timeouts on the autopkgtest side.

And finally the regular deb is build as part of the autopkgtest
run. Here the unittests are already run so we can skip the
tests/unit part of spread.