cmd/snap-update-ns: use x-snapd.{synthetic,needed-by} in practice
The mount diff/update mechanism had a simple idea on how to handle
updates when faced with synthesized mount entries. The synthetic mount
entries were tagged with the x-snapd.synthetic and
x-snapd.needed-by=something tags and the update mechanism would
understand them and keep the synthetic entires around for as long as the
real entry that referred to them was kept.

The problem is this was never really done in practice. The mimic code
never generated the tags. Thus mimic code would cause issues when
refreshing snaps.

This patch makes the prepareWritableMimic emit the right tags and
updates the tests to match. The spread test for layout is tweaked
to run twice.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>