snap-confine: fallback to /lib/udev/snappy-app-dev if the core is older
There is code in snap-confine that runs inside the core snap and runs
our udev helper. We have two paths for the udev helper, the old path
/lib/udev/snappy-appy-dev which we no longer use but may still be
present if the user has an older version of the core snap installed.
And the new path /usr/lib/snapd/snap-device-helper.

This PR changes snap-confine to try /usr/lib/snapd/snap-device-helper
first but fallback to /lib/udev/snappy-app-dev if the new one is not
available.