cmd/snap-update-ns: don't trespass on host filesystem
This patch teaches snap-update-ns not to create files, directories and
symlinks in places that directly affect the host. One example of such
place includes the /etc directory which is writable (thus doesn't
trigger the creation of a writable mimic). If a layout puts a new file,
directory or symlink there then the host will observe an empty
directory, an empty file or an actual symlink.

Because this is highly undesirable the code that handles read-only
filesystem is generalized to handle "trespassing" in the exact same way,
namely by constructing a mimic at the appropriate location. In the /etc
example above, the real etc would be replaced by a mimic and the newly
created file would not be present in the host filesystem anymore.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>