snap: make `snap run` look at the system-key for security profiles
many: make `snap run` look at the system-key for security profiles

If the system-key for the security profiles does not match, wait and
eventually timeout if snapd is not generating the profiles we need.

Without it we may have the following scenario:

1. snapd gets refreshed and snaps need updated security profiles
   to work (e.g. because snap-confine needs new generated
   snap-update-ns profiles on disk)
2. The system reboots to start the new snapd. At this point
   the old security profiles are on disk (because the new
   snapd did not run yet) and nothing new is there yet.
3. Snaps that run as daemon get started during boot by systemd
   (e.g. network-manager). This may happen before snapd had a
   chance to refresh the security profiles.
4. Because the security profiles are for the old version of
   the snaps that run before snapd starts and generates the
   missing security profiles will fail to start. For e.g.
   network-manager this is of course catastrophic as it
   leaves the machine without networking.

To prevent this, in step(4) we have this wait-for-snapd
step to ensure the expected profiles are on disk.

This PR also switches the system-key to json to make the
parsing a lot quicker.

It also introduces a new version key into the system-key.
This version key is important because there is a race when
a) the sytem-key gets new inputs
b) old `snap run` is executed and slow to load
c) new snapd with new inputs run at exactly the same time and
   write a new system-key to disk
In this case (b) will have calculated a profile without the new
inputs and eventually timeout. The new code will just continue
if there is a version mismatch (which is fine because it means
new profiles are on disk so no need to wait).