cmd/snap-confine: nvidia: preserve globbed file prefix
When symlinking libgl files from hostfs we discard the path prefix inside
source_dir of a globbed file and overwrite already symlinked files if those are
created before.

For example, nvidia globs:
  libnvidia-tls.so*
  tls/libnvidia-tls.so*

Produce the following matches:

  ${source_dir}/libnvidia-tls.so
  ${source_dir}/libnvidia-tls.so.390.42
  ${source_dir}/tls/libnvidia-tls.so
  ${source_dir}/tls/libnvidia-tls.so.390.42

When the files under ${source_dir}/tls are processed, eg.
${source_dir}/tls/libnvidia-tls.so.390.42 the symlink we create will
overwrite ${libgl_dir}/libnvidia-tls.so.390.42, thus shadowing the original
library that was present at this path.

The patch attempts to fix this behavior and mirrors the source_dir globs, thus
the following symlinks are created:

  ${libgl_dir}/libnvidia-tls.so.390.42 -> ${source_dir}/libnvidia-tls.so.390.42
  ${libgl_dir}/tls/libnvidia-tls.so.390.42 -> ${source_dir}/tsls/libnvidia-tls.so.390.42

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>