overlord/configstate: change how ssh is stopped/started
The ssh.service has an alias of sshd.service and so we must operate on
the main file ssh.service for enable/disable operations as aliases do
not exist when the main unit is disabled.

Due to some unfortunate use of writable paths we cannot mask ssh.service
and have that unit stay gone on reboot. To work around that use the fact
that the unit is conditional of non-existence of /etc/ssh/sshd_not_to_be_run

The patch also unmasks ssh{,d}.service in case it was masked by earlier
snapd.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>